policy-sim/requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0
sentry-sdk[fastapi]>=2.0.0
edge-tts>=7.0.0
